requirements.txt
aiofiles==23.2.1
fastapi==0.110.1
Jinja2==3.1.3
openai==1.19.0
python-dotenv==1.0.1
python-multipart==0.0.9
uvicorn==0.29.0
websockets==12.0
